Fix action item description not being a markdown editor
